Fixes #4551 - GCE provisioning support
Enable provisioning of VMs through Google Compute Engine. Volume-wise,
this is currently limited to creating a VM with an attached disk that
contains the image specified. Future enhancements should include
choosing any available disks to auto-attach the VM and not force the
user to create a new one through Foreman.