fixes #13111 - use negated Capybara lookups to wait for JS handlers
Various integration tests failed intermittently as they didn't wait for
asynchronous JS to run when clicking buttons.  Instances of refuting
existence of an element after a click were changed to negated Capybara
lookups which wait until the element is removed, instead of failing
immediately when finding an element that hadn't yet been removed.

Assertions on the value of current_path are changed to the built-in
assertion in Capybara 2.5 which waits for redirections to occur.

Added the capybara-screenshot gem to capture integration test failures
and store them in tmp/capybara/.