fixes #15751 - replace spork with spring in development
To make use of it, run bin/rails and bin/rake rather than plain
rails or rake commands, and run `spring status` or `spring stop` to
check and stop current background processes.

Test log removal has moved to the logging framework as the file would
not be recreated after being deleted in the test helper, since the
Rails environment and logger are not reloaded on each `rake test`.