Add ParameterFilters outline
The basic idea is a class that stores permit data per model. It will
be stored as some kind of per-model class-level variable eventually,
or in a controller concern (per model). It should be easily accessible
to plugins so they can call #permit, or the concern or whatever's
instantiating it could call #permit for all known model attributes and
then could call something on the plugin interface so they can then
extend it further.

It allows a series of blocks to be registered that call #permit on the
controller parameters object based on the type of request (API, UI).

The same object should be reusable for nested models, if that's added
to the block context. Attributes ought to be permitted by default in the
UI and API if they're known, but blocked by default on nested models.