Fix setup of JSON parse error middleware
The ParamsParser middleware is no longer loaded by default in Rails 5,
and params parsing is delayed until it's required allowing the JSON
verification middleware to be placed at any point.

For backporting it's probably best to make this conditional on the Rails
version - else ParamsParser _could_ be enabled on Rails 5, but it'll
issue an unnecessary extra warning.

Required for rails5.
Can't be backported to develop as-is, needs conditionals.