fixes #10982 - configure RestClient log to 'proxy' logger
Request/response information is sent to the proxy logger, showing the
URLs called, headers, request body etc.

The X-Request-ID header is now also set in the proxy requests to the
same session ID that is logged by default in Foreman. The ID is now a
digest of the original session ID to prevent the user's session ID being
hijacked from the proxy request data.