Fix TaxHost caching in test by reloading taxonomy
Taxonomy tests for used IDs were caching the list of associated hosts
through TaxHost from when the taxonomy was created, before the host was
assigned. Reloading the taxonomy after host assignment ensures the hosts
relation (which is also cached inside TaxHost) shows the current state.

    LocationTest.test_0008_it should return array of used ids by hosts
    Expected: []
      Actual: [334344675] (Minitest::Assertion)
    /var/lib/workspace/workspace/test_develop_pr_core@2/database/postgresql/label/fast/ruby/2.3/test/unit/location_test.rb:84

Can be backported to develop.
Required for rails5.