Fix cached nested usergroups in test setup
A usergroup membership test was failing with:

    UsergroupMemberTest#test_0014_diamond-with-tail usergroups, one of way is removed [test/unit/usergroup_member_test.rb:241]:
    Expected [#<Role id: 7, name: "Default role", builtin: 2>] to include #<Role id: 13, name: "um_role", builtin: 0>.

When the role was created in this test, it's meant to recurse into the
usergroups that are members of `@a` and create UserRole objects for any
users that inherited the role. The users assigned to `um_d` were
previously cached in the copy of `um_d` inside `@a`, so didn't show
`@user1`'s membership. Reloading all of `@a` ensures caches are empty.

Can be backported to develop.
Required for rails5.