Explicitly reset format on subsequent functional requests
    HostsControllerTest#test_0037_if only authorize_login_delegation is set, REMOTE_USER should be
            ignored for API requests [test/controllers/hosts_controller_test.rb:617]:
    Expected response to be a <2XX: success>, but was a <401: Unauthorized>

When making a JSON format request followed by a request with no explicit
format, Rails 5 is now memorising that it's requesting application/json
instead of resetting to HTML or similar. Almost certainly a Rails bug.

Explicitly setting the expected format type works around it.

Can be backported to develop.
Required for rails5.