Re-enable deep_munge for security
The deep_munge option should remain enabled to match Rails defaults,
which means nulls are stripped from arrays by default.

Apart from fixing the conversion of `[]` to `nil` in Rails 4.2, there
seems to be no other reason to disable it, so leaving it at the default
under Rails 5 is likely to be best - in case there are future security
issues or bugs it prevents.

Cannot be backported to develop.
Not required for rails5.