fixes #18500 - replace quiet_assets with sprockets-rails flag
quiet_assets was merged into sprockets-rails 3.1+ and is no longer
maintained for Rails 5, so it can be removed. Min version hasn't been
increased as it only affects development log verbosity and isn't a hard
dependency. Older versions used in packages will work fine.

The logging gem deliberately doesn't implement the #silence method, so
this is wrapped to fully match the Rails API and make the feature work.