Replace alias_method_chain with Module prepend
Deprecated in 5.0 and now removed in 5.1. Some instances of classes
overwriting existing methods can be handled with `super`, other concerns
or modules are changed to use prepend instead of include. (Note, no
ActiveSupport::Concern support for prepends, so load class methods the
pure Ruby way.)

Required for rails51.
Can be backported to develop.